Added addition attributes; fixed bug in calendar output
Tidied up  code where needed.  Added a date_start attribute to the Event_List and Event_Card blocks, with code changes required for it in the php.  Fixed an error in the display of calendar, and made sure we remove days with no events from the small screen event list rendering of the calendar. Edited the readme.txt to reflect all changes and to give some doc on the attributes.